requirements.txt
openwakeword>=0.6.0
sounddevice>=0.4.6
numpy>=1.24.0
RealtimeSTT>=0.1.0
onnxruntime-cpu>=1.15.0
pyttsx3>=2.90
google-generativeai>=0.3.0
aiohttp>=3.8.0
psutil>=5.9.0 